Add files via upload
This PowerShell script solves a problem that I've run into at work in the past few days; there is a large number of USB-powered card readers being rolled out soon, and as I've found out, they require a setting in the registry that exists on our mini-model desktops to be set to false so that they aren't suspended from power by the computer. This script searches the known registry for the known VID and grabs all subkeys and puts them in an array, then appends the key path that holds the entry which needs changed to the paths that the array contains (amount of paths varies depending on the number of different devices (same VID) that has been plugged into the computer). It then changes the "EnhancedPowerManagement" key entry from 1 to 0 or, if the entry doesn't exist, it adds it and sets it to 0. However, if the entry does not exist, this script is not necessary; it only adds it if it doesn't exist because I used the "Set-ItemProperty" cmdlet.

Includes the original PowerShell script and the script wrapped with batch to make an executable .cmd file.